+++ src/spot/lib/Model/SorOrderRequest.php
@@ -73,7 +73,7 @@ class SorOrderRequest implements ModelInterface, \ArrayAccess, \JsonSerializable
         'icebergQty' => 'float',
         'newOrderRespType' => '\Binance\Client\Spot\Model\NewOrderRespType',
         'selfTradePreventionMode' => '\Binance\Client\Spot\Model\SelfTradePreventionMode',
-        'recvWindow' => 'int',
+        'recvWindow' => 'float',
     ];
 
     /**
@@ -98,7 +98,7 @@ class SorOrderRequest implements ModelInterface, \ArrayAccess, \JsonSerializable
         'icebergQty' => 'float',
         'newOrderRespType' => null,
         'selfTradePreventionMode' => null,
-        'recvWindow' => 'int64',
+        'recvWindow' => 'float',
     ];
 
     /**
@@ -693,7 +693,7 @@ class SorOrderRequest implements ModelInterface, \ArrayAccess, \JsonSerializable
     /**
      * Gets recvWindow.
      *
-     * @return null|int
+     * @return null|float
      */
     public function getRecvWindow()
     {
@@ -703,7 +703,7 @@ class SorOrderRequest implements ModelInterface, \ArrayAccess, \JsonSerializable
     /**
      * Sets recvWindow.
      *
-     * @param null|int $recvWindow recvWindow
+     * @param null|float $recvWindow recvWindow
      *
      * @return self
      */
